+++ whereexpr/src/cond_parser/modifiers.rs
@@ -24,14 +24,14 @@ pub(crate) fn parse(text: &str) -> Result<(Modifiers, usize), Error> {
     // find opening {
     let start = trimmed
         .rfind('{')
-        .ok_or_else(|| Error::UnmatchedModifierBracket(close_pos as u16, (close_pos + 1) as u16, text.to_string()))?;
+        .ok_or_else(|| Error::UnmatchedModifierBracket(close_pos as u32, (close_pos + 1) as u32, text.to_string()))?;
 
     // extract content between { and }
     let content = &trimmed[start + 1..close_pos];
 
     // empty modifier block e.g. {}
     if content.trim().is_empty() {
-        return Err(Error::EmptyModifierBlock(start as u16, (close_pos+1) as u16, text.to_string()));
+        return Err(Error::EmptyModifierBlock(start as u32, (close_pos+1) as u32, text.to_string()));
     }
 
     // parse modifiers tracking position of each one
@@ -49,7 +49,7 @@ pub(crate) fn parse(text: &str) -> Result<(Modifiers, usize), Error> {
         match trimmed_modifier {
             "ignore-case" => modifiers.ignore_case = true,
             _ => {
-                return Err(Error::UnknownModifier(modifier_start as u16, modifier_end as u16, text.to_string()));
+                return Err(Error::UnknownModifier(modifier_start as u32, modifier_end as u32, text.to_string()));
             }
         }
